refactor(parser): replace the regex Mermaid front-end with a lark grammar
The Mermaid + %%metro front-end was a line-by-line regex dispatcher: six
order-sensitive node-shape patterns, an edge pattern, and an if-elif statement
classifier, plus per-directive string slicing. Replace it with a declarative
lark grammar and carry that structure through the rest of the parse layer.

- grammar.py: the grammar, typed statement dataclasses, and a transformer that
  flattens the parse tree into ordered, already-normalised statements;
  parse_statements() wraps the parse (and surfaces parse errors as ValueError).
  All node shapes collapse to one SHAPE terminal (the model records only id +
  label), and edge endpoints may carry an inline shape.
- directives.py: directive bodies dispatch through a _GLOBAL_DIRECTIVE_HANDLERS
  registry keyed by exact name (so a key can't shadow a longer one), with the
  section-scoped entry/exit/direction and file/files/dir icons handled
  alongside. Every handler warns-and-ignores unusable payload through one
  _warn_directive mechanism, rather than some warning and some silently
  returning.
- resolve.py: the post-parse section/port/junction rewrites, unchanged.
- mermaid.py: the public parse_metro_mermaid driver, split into
  _apply_statements / _finalize_graph / _apply_pending_metadata.

Leniency is now explicit and uniform: unrecognised lines and unknown
directives warn rather than vanishing silently; syntax stays lenient while
semantic validity remains a separate, stricter phase. Parsed MetroGraph output
is byte-identical across the fixture corpus. Adds the lark runtime dependency.
