+++ pyproject.toml
@@ -23,6 +23,7 @@ classifiers = [
 dependencies = [
     "click>=8.0",
     "drawsvg>=2.0",
+    "lark>=1.1",
     "networkx>=3.0",
     "pillow>=9.0",
 ]
